PicoTTS added - compiled from source (#142)
* add picotts library to docker image

* Added picotts build from src

PicoTTS is a locally installed Text-To-Speach engine. It is supported by home assistant and should be installed in the docker container by default. Since it is not supported by alpiine 3.12, it must be compiled on the fly.

* Minor fixes

The build should be reproducable - therefore the current commit hash should be used and checked out
The packages must be removed afterwards - so a virtual alias should be used
The package-cache must not be used
The clone must be removed afterwards

* Added value for picotts_hash

In order to produce a better pipeline build, the picotts_hash shall be taken form the build.json file.

* Some tweaks

Using no longer a default value for picotts hash
Removed git from .build-dependencies (Telldus won't be available afterwards
Swapped removal of build-dependencies and source directory

* Fixed whitespace error

* Fixed missing runtime-dependency for popt

* Applied build argument correctly

* Hadolint failure

* Resolving review comments

Fixing format of installing packages and filesystem navigation

* Moved PICO_TTS hash to args and merged with current master

* Disable static archives on picoTTS

Co-authored-by: kngharv <kngharv@gmail.com>