Add Renovate config for home-assistant/docker-base updates
Make sure the image stays up-to-date by updating the pinned version of
the base image by the Renovate bot. The GH release tag is used as the
source of truth, with 1 hour cooldown, which should be now plenty for
the images to be available in the container registry (the build should
take less than 30 minutes on native builders).