Rebuild wheels also when workflow changes (#375)
* Rebuild wheels also when workflow changes

In #373 we missed the breaking change because because the requirements
were not changed and the wheels build happened only on release. Rebuild
wheels also when the builder workflow changes.

* Don't run changed-files action on release builds

The changed-files actions fails on release build because it only has
shallow clone of the repository. Per [1] we'd either need to increase
fetch-depth or persist credentials for git. Since we we always run the
depending step on release build, we don't have to do any of this and
simply skip checking for the changed files.

[1] https://github.com/tj-actions/changed-files?tab=readme-ov-file#usage-